Изменен тип для идентификатора в IOrdersRepository (Orders)

+++ Source/Backend/Csharp/Orders/Marketplace.Orders.Application/Common/Exceptions/OrderNotFoundException.cs
@@ -4,6 +4,6 @@
 /// Заказ не найден
 /// </summary>
 /// <param name="id">Идентификатор</param>
-public class OrderNotFoundException(int id)
+public class OrderNotFoundException(long id)
     : Exception($"Заказ с идентификатором {id} не найден")
 { }
